requirements.txt
anthropic>=0.39.0
fastapi>=0.104.0
uvicorn>=0.24.0
click>=8.1.0
python-dotenv>=1.0.0
